refactor(studio): one StatCard, in components/
The PR created `features/reports/stat-card.tsx` while `fanout-panel.tsx`, five
files away in the same directory, kept its own — a card created to be shared
with a duplicate as its neighbour. `fanout-panel`'s props are a strict subset
(`{label, testId, value}` against `{chart?, footer?, label, testId?, value}`)
and the markup differed only by a `justify-between` and a flex wrapper that
have no effect without a chart or footer, so it swaps cleanly.

Moved to `src/components/`, the studio's primitives home, because three
features render this anatomy and none of them owns it. Its props are `readonly`
now, matching the interfaces around them.

Two things deliberately NOT unified, recorded so the next reader doesn't have
to re-derive them:

- **`home-panel`'s `StatCard` is a different component** that happens to share
  the noun — it carries a delta, a trend line, and a unit, and calls `useT()`.
  Folding it in would be a redesign, not a de-duplication. Noted in the shared
  card's docblock.
- **The three remaining `formatMs` variants are three different formats**, not
  three copies: `fanout-panel` renders `1234 ms`, `home-panel` `24ms`/`1.2s`,
  `function-stats` `9.5ms`/`1.23s`, and `metrics-format.formatLatency`
  `400μs`/`1.00s`. Unifying them changes what operators read on four screens,
  which is a product call and not something to slip into a refactor that claims
  no behaviour change.

958 tests pass.
